requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
appwrite
mem0ai
tavily-python
langgraph
langchain-google-genai
langchain
langchain-core
python-dotenv
httpx
PyPDF2
python-docx
pydantic-settings
chromadb
pydantic[email]